Refactored CLI Fixed path building in fluent-token parser
