+++ aiogram_i18n/managers/base.py
@@ -1,5 +1,7 @@
+from __future__ import annotations
+
 from abc import abstractmethod, ABC
-from typing import Dict, Any, TYPE_CHECKING, Awaitable, Callable, Optional
+from typing import Any, Awaitable, Callable, Dict, Optional, TYPE_CHECKING
 
 from aiogram.dispatcher.event.handler import CallableMixin
 from aiogram.types import TelegramObject
@@ -7,11 +9,11 @@ from aiogram.types import TelegramObject
 
 class BaseManager(ABC):
     default_locale: Optional[str]
-    set_locale_mixin: CallableMixin
+    locale_setter: LocaleSetter
 
     def __init__(self, default_locale: Optional[str] = None) -> None:
         self.default_locale = default_locale
-        self.set_locale_mixin = CallableMixin(self.set_locale)
+        self.locale_setter = LocaleSetter(self)
 
     @abstractmethod
     async def get_locale(self, event: TelegramObject, data: Dict[str, Any]) -> str:
@@ -24,3 +26,11 @@ class BaseManager(ABC):
         @abstractmethod
         async def set_locale(self, *args: Any, **kwargs: Any) -> None:
             ...
+
+
+class LocaleSetter(CallableMixin):
+    def __init__(self, manager: BaseManager) -> None:
+        super().__init__(callback=manager.set_locale)
+
+    async def __call__(self, locale: str, /, **kwargs: Any) -> Any:
+        return await self.call(locale, **kwargs)
