+++ aiogram_i18n/utils/cli/__init__.py
@@ -1,0 +1,13 @@
+
+from .base import main
+from .extract import extract
+from .multiple_extract import multiple_extract
+from .stub import stub
+
+
+__all__ = [
+    "main",
+    "extract",
+    "multiple_extract",
+    "stub"
+]
