+++ aiogram_i18n/utils/cli/base.py
@@ -1,0 +1,7 @@
+
+import click
+
+
+@click.group(name="main")
+def main() -> None:
+    ...
